% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_magic.R
\name{do_magic}
\alias{do_magic}
\title{Multiples as values by two, except 42, which stays 42}
\usage{
do_magic(x)
}
\arguments{
\item{x}{input, must be numeric}
}
\value{
magicified output
}
\description{
Multiples as values by two, except 42, which stays 42
}

